unblockjam: (refactor) move block drag logic to a new file, preparatory change of exit visuals
